#s9-22 Правка типа для StoreInterval

+++ internal/server/config/config.go
@@ -4,7 +4,7 @@ package config
 type Config struct {
 	Host            string `valid:"minstringlength(3)"`
 	LogLevel        string `valid:"in(debug|info|warn|error|fatal)"`
-	StoreInterval   int    `valid:"range(1|999)"`
+	StoreInterval   int64  `valid:"range(1|999)"`
 	RepeatIntervals []int
 	FileStoragePath string `valid:"minstringlength(3)"`
 	Restore         bool
