+++ internal/server/services/flagsparser/flags_parser.go
@@ -12,7 +12,7 @@ type flagsParser struct {
 // Parse парсит аргументы командной строки и наполняет конфигурацию
 func (p flagsParser) Parse(c *config.Config) error {
 	fl.StringVar(&c.Host, "a", c.Host, "Net address host:port")
-	fl.IntVar(&c.StoreInterval, "i", c.StoreInterval, "Store interval")
+	fl.Int64Var(&c.StoreInterval, "i", c.StoreInterval, "Store interval")
 	fl.StringVar(&c.FileStoragePath, "f", c.FileStoragePath, "File storage path (full)")
 	fl.BoolVar(&c.Restore, "r", c.Restore, "Restore values on start")
 	fl.StringVar(&c.DatabaseDSN, "d", c.DatabaseDSN, "Database DSN")
